This was a bad bug
For images with few colours (THIS IS FINE) the algorithm would fail due
to having clusters with 0 items. Changed a bit the code so that the
number of clusters is adjusted during iteration. A better way would
have been to adjust initial centroids to be "different enough" but
that risks infinite loops anyway. This at least works "as expected".
Press R if you don't see enough colours!